+++ src/backend/ci/core/environment/biz-environment/src/main/kotlin/com/tencent/devops/environment/service/NodeService.kt
@@ -388,8 +388,4 @@ class NodeService @Autowired constructor(
             records = nodeInfos.map { NodeStringIdUtils.getNodeBaseInfo(it) }
         )
     }
-
-    fun getOs(userId: String, projectId: String, agentId: String): String {
-        return nodeDao.get(dslContext, projectId, HashUtil.decodeIdToLong(agentId))?.osName ?: "linux"
-    }
 }
